Move stuff to the physical/mock folder
