+++ src/query/v2/physical/mock/mock.hpp
@@ -1,0 +1,116 @@
+// Copyright 2022 Memgraph Ltd.
+//
+// Use of this software is governed by the Business Source License
+// included in the file licenses/BSL.txt; by using this file, you agree to be bound by the terms of the Business Source
+// License, and you may not use this file except in compliance with the Business Source License.
+//
+// As of the Change Date specified in that file, in accordance with
+// the Business Source License, use of this software will be governed
+// by the Apache License, Version 2.0, included in the file
+// licenses/APL.txt.
+
+#pragma once
+
+#include <vector>
+
+#include "query/v2/physical/mock/frame.hpp"
+#include "query/v2/physical/multiframe.hpp"
+#include "query/v2/physical/physical.hpp"
+#include "utils/logging.hpp"
+
+namespace memgraph::query::v2::physical::mock {
+
+using TFrame = Frame;
+using TDataPool = physical::multiframe::MPMCMultiframeFCFSPool;
+using TPhysicalOperator = physical::PhysicalOperator<TDataPool>;
+using TPhysicalOperatorPtr = std::shared_ptr<TPhysicalOperator>;
+using TOnceOperator = physical::OncePhysicalOperator<TDataPool>;
+template <typename TDataFun>
+using TScanAllOperator = physical::ScanAllPhysicalOperator<TDataFun, TDataPool>;
+using TProduceOperator = physical::ProducePhysicalOperator<TDataPool>;
+using TExecutionContext = physical::ExecutionContext;
+
+enum class OpType { Once, ScanAll, Produce };
+inline std::ostream &operator<<(std::ostream &os, const OpType &op_type) {
+  switch (op_type) {
+    case OpType::Once:
+      os << "Once";
+      break;
+    case OpType::ScanAll:
+      os << "ScanAll";
+      break;
+    case OpType::Produce:
+      os << "Produce";
+      break;
+  }
+  return os;
+}
+
+constexpr static int SCANALL_ELEMS_POS = 0;
+struct Op {
+  OpType type;
+  std::vector<int> props;
+};
+
+inline void LogOps(const std::vector<Op> &ops) {
+  for (const auto &op : ops) {
+    if (op.type == OpType::ScanAll) {
+      SPDLOG_INFO("{} elems: {}", op.type, op.props[SCANALL_ELEMS_POS]);
+    } else {
+      SPDLOG_INFO("{}", op.type);
+    }
+  }
+}
+
+inline TPhysicalOperatorPtr MakeOnce(int pool_size, int mf_size) {
+  auto data_pool = std::make_unique<TDataPool>(pool_size, mf_size);
+  return std::make_shared<TOnceOperator>(TOnceOperator("Physical Once", std::move(data_pool)));
+}
+
+inline TPhysicalOperatorPtr MakeScanAll(int pool_size, int mf_size, int scan_all_elems) {
+  auto data_fun = [scan_all_elems](TDataPool::TMultiframe &mf, TExecutionContext &) {
+    std::vector<TFrame> frames;
+    for (int i = 0; i < scan_all_elems; ++i) {
+      for (int j = 0; j < mf.Data().size(); ++j) {
+        frames.push_back(TFrame{});
+      }
+    }
+    return frames;
+  };
+  auto data_pool = std::make_unique<TDataPool>(pool_size, mf_size);
+  return std::make_shared<TScanAllOperator<decltype(data_fun)>>(
+      TScanAllOperator<decltype(data_fun)>("Physical ScanAll", std::move(data_fun), std::move(data_pool)));
+}
+
+inline TPhysicalOperatorPtr MakeProduce(int pool_size, int mf_size) {
+  auto data_pool = std::make_unique<TDataPool>(pool_size, mf_size);
+  return std::make_shared<TProduceOperator>(TProduceOperator("Physical Produce", std::move(data_pool)));
+}
+
+inline TPhysicalOperatorPtr MakePlan(const std::vector<Op> &ops, int pool_size, int mf_size) {
+  TPhysicalOperatorPtr plan = nullptr;
+  auto current = plan;
+  for (const auto &op : ops) {
+    if (op.type == OpType::Once) {
+      auto once = MakeOnce(pool_size, mf_size);
+      current->AddChild(once);
+      current = once;
+
+    } else if (op.type == OpType::ScanAll) {
+      auto scan_all = MakeScanAll(pool_size, mf_size, op.props[SCANALL_ELEMS_POS]);
+      current->AddChild(scan_all);
+      current = scan_all;
+
+    } else if (op.type == OpType::Produce) {
+      auto produce = MakeProduce(pool_size, mf_size);
+      plan = produce;  // top level operator
+      current = plan;
+
+    } else {
+      SPDLOG_ERROR("Unknown operator {}", op.type);
+    }
+  }
+  return plan;
+}
+
+}  // namespace memgraph::query::v2::physical::mock
